utils: Optional pygit2 implementation for git_mtime()
Instead of spawning a subprocess for every check, use pygit2 to
calculate mtimes.

This approach gives some speedups.  Benchmarks for 100 calls to
*_mtime() for 'docs/faq.rst' in the current Pelican source:

  use_last_modification  _pygit_mtime  subprocess_git_mtime
  =====================  ============  ====================
  True                   2 s           10 s
  False                  6 s           10 s

The subprocess times are constant because the subprocess
implementation currently reads in the whole file log.  The pygit2
times depend on how far down the list of commits you have to seek.
Without a way to walk the log for a single file (written in C in
pygit2), there's probably no way around this that scales across a
whole working directory.  I don't think these gains are worth the
added code complexity.